refactor: upgrade debug and only support new node versions
BREAKING CHANGE: node >= 4.0.0
