--- .npmignore
@@ -1,9 +1,0 @@
-test/
-coverage.html
-Makefile
-.travis.yml
-logo.png
-.jshintignore
-.jshintrc
-.gitingore
-coverage/
