+++ .travis.yml
@@ -1,4 +1,19 @@
+sudo: false
 language: node_js
+notifications:
+  email: false
 node_js:
-  - '0.11'
-  - '0.10'
+  - '8'
+  - '6'
+  - '4'
+install:
+  - npm i npminstall && npminstall
+script:
+  - npm run ci
+after_script:
+  - npminstall codecov && codecov
+after_success:
+  - npm run semantic-release
+branches:
+  except:
+    - /^v\d+\.\d+\.\d+$/
