--- AUTHORS
@@ -1,3 +1,0 @@
-fengmk2 <fengmk2@gmail.com> (https://github.com/fengmk2)
-popomore <sakura9515@gmail.com> (https://github.com/popomore)
-良田 <chenhao.lch@alibaba-inc.com>
