+++ LICENSE.txt
@@ -1,7 +1,7 @@
 This software is licensed under the MIT License.
 
+Copyright (C) 2015 - 2017 node-modules and other contributors
 Copyright (C) 2014 fengmk2 <fengmk2@gmail.com> and other contributors
-Copyright (C) 2015 node-modules and other contributors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
