--- Makefile
@@ -1,37 +1,0 @@
-TESTS = test/*.test.js
-REPORTER = spec
-TIMEOUT = 10000
-MOCHA_OPTS =
-
-install:
-	@npm install
-
-jshint:
-	@./node_modules/.bin/jshint .
-
-test:
-	@NODE_ENV=test ./node_modules/.bin/mocha \
-		--harmony \
-		--reporter $(REPORTER) \
-		--timeout $(TIMEOUT) \
-		$(MOCHA_OPTS) \
-		$(TESTS)
-
-test-cov cov:
-	@NODE_ENV=test node --harmony \
-		node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha \
-		-- -u exports \
-		--reporter $(REPORTER) \
-		--timeout $(TIMEOUT) \
-		$(MOCHA_OPTS) \
-		$(TESTS)
-	@./node_modules/.bin/cov coverage
-
-autod: install
-	@./node_modules/.bin/autod -w
-	@$(MAKE) install
-
-contributors:
-	@./node_modules/.bin/contributors -f plain -o AUTHORS
-
-.PHONY: test
