+++ package.json
@@ -1,39 +1,45 @@
 {
   "name": "speeds",
-  "version": "0.0.1",
+  "version": "0.0.0-development",
   "description": "Speeds controllable stream. Uou can control the downstream speed on your own.",
   "main": "index.js",
+  "files": [
+    "index.js"
+  ],
   "scripts": {
-    "test": "make test"
+    "test": "mocha -t 10000 test/*.test.js",
+    "semantic-release": "semantic-release pre && npm publish && semantic-release post",
+    "autod": "autod --prefix '^' -w"
   },
   "dependencies": {
-    "debug": "~2.1.3"
+    "debug": "^2.6.8"
   },
   "devDependencies": {
     "autod": "*",
     "contributors": "*",
-    "cov": "*",
     "istanbul": "*",
-    "jshint": "*",
     "mocha": "*",
-    "should": "3.1.3"
+    "semantic-release": "^6.3.6",
+    "should": "^11.2.1"
   },
-  "homepage": "https://github.com/fengmk2/speeds",
   "repository": {
     "type": "git",
-    "url": "git://github.com/fengmk2/speeds.git",
-    "web": "https://github.com/fengmk2/speeds"
+    "url": "https://github.com/node-modules/speeds.git"
   },
   "bugs": {
-    "url": "https://github.com/fengmk2/speeds/issues",
+    "url": "https://github.com/node-modules/speeds/issues",
     "email": "fengmk2@gmail.com"
   },
   "keywords": [
-    "speeds", "stream", "stream2", "Transform", "downstream"
+    "speeds",
+    "stream",
+    "stream2",
+    "Transform",
+    "downstream"
   ],
   "engines": {
-    "node": ">= 0.10.0"
+    "node": ">= 4.0.0"
   },
-  "author": "fengmk2 <fengmk2@gmail.com> (http://fengmk2.github.com)",
+  "author": "fengmk2 <fengmk2@gmail.com> (https://fengmk2.com)",
   "license": "MIT"
 }
