+++ test/demo.js
@@ -1,18 +1,4 @@
-/**!
- * speeds - test/demo.js
- *
- * Copyright(c) fengmk2 and other contributors.
- * MIT Licensed
- *
- * Authors:
- *   fengmk2 <fengmk2@gmail.com> (http://fengmk2.github.com)
- */
-
 'use strict';
-
-/**
- * Module dependencies.
- */
 
 var speeds = require('../');
 var fs = require('fs');
