+++ test/speeds.test.js
@@ -1,18 +1,4 @@
-/**!
- * speeds - test/speeds.test.js
- *
- * Copyright(c) 2014 fengmk2 and other contributors.
- * MIT Licensed
- *
- * Authors:
- *   fengmk2 <fengmk2@gmail.com> (http://fengmk2.github.com)
- */
-
-"use strict";
-
-/**
- * Module dependencies.
- */
+'use strict';
 
 var should = require('should');
 var fs = require('fs');
